+++ src/PSNtfy/public/Receive-NtfyPush.ps1
@@ -111,9 +111,7 @@ function Receive-NtfyPush {
     }
 
     # Join Headers into Payload
-    if($Headers.Count -gt 0) {
-        $Payload["Headers"] = $Headers
-    }
+    Add-ObjectPropSafe -Object $Payload -Key "Headers" -Value $Headers
 
     # Receive Response
     $response = Invoke-RestMethod @Payload